artifacts/backend/requirements.txt
# Backend Python dependencies
fastapi
uvicorn
pydantic
python-dotenv
openai
psycopg2-binary
pypdf
python-docx
sqlalchemy
